Nimbus/cloud-run-function/requirements.txt
functions-framework==3.*
pandas
requests
google-cloud-bigquery
db-dtypes
pyarrow